requirements.txt
pipecat-ai[daily,openai,sarvam]
python-dotenv
loguru
asyncpg
python-dateutil